+++ packages/adapter-ejs/test/ejs_adapter_test.js
@@ -34,7 +34,7 @@ describe('EJS adapter', () => {
 
   describe('#filesForComponent', () => {
     it('should return the component file', () => {
-      const files = Adapter.filesForComponent('button')
+      const files = Adapter.filesForComponent({}, 'button')
 
       assert.strictEqual(files.length, 1)
       assert.strictEqual(files[0].basename, 'button.ejs')
@@ -43,7 +43,7 @@ describe('EJS adapter', () => {
 
   describe('#filesForVariant', () => {
     it('should return the variant file', () => {
-      const files = Adapter.filesForVariant('button', 'primary')
+      const files = Adapter.filesForVariant({}, 'button', 'primary')
 
       assert.strictEqual(files.length, 1)
       assert.strictEqual(files[0].basename, 'primary.ejs')