+++ packages/adapter-marko/src/index.js
@@ -35,7 +35,7 @@ async function render (options, filePath, data = {}) {
   })
 }
 
-function filesForComponent (componentName) {
+function filesForComponent (options, componentName) {
   return [
     {
       basename: `${componentName}.marko`,
@@ -44,7 +44,7 @@ function filesForComponent (componentName) {
   ]
 }
 
-function filesForVariant (componentName, variantName) {
+function filesForVariant (options, componentName, variantName) {
   return [
     {
       basename: `${variantName}.marko`,