+++ packages/adapter-pug/test/pug_adapter_test.js
@@ -59,7 +59,7 @@ describe('Pug adapter', () => {
 
   describe('#filesForComponent', () => {
     it('should return the component file', () => {
-      const files = Adapter.filesForComponent('button')
+      const files = Adapter.filesForComponent({}, 'button')
 
       assert.strictEqual(files.length, 1)
       assert.strictEqual(files[0].basename, 'button.pug')
@@ -68,7 +68,7 @@ describe('Pug adapter', () => {
 
   describe('#filesForVariant', () => {
     it('should return the variant file', () => {
-      const files = Adapter.filesForVariant('button', 'primary')
+      const files = Adapter.filesForVariant({}, 'button', 'primary')
 
       assert.strictEqual(files.length, 1)
       assert.strictEqual(files[0].basename, 'primary.pug')