+++ packages/adapter-react/test/react_adapter_test.js
@@ -224,7 +224,7 @@ describe('React adapter', () => {
 
   describe('#filesForComponent', () => {
     it('should return the component file', () => {
-      const files = Adapter.filesForComponent('button')
+      const files = Adapter.filesForComponent({}, 'button')
 
       assert.strictEqual(files.length, 1)
       assert.strictEqual(files[0].basename, 'Button.jsx')
@@ -233,7 +233,7 @@ describe('React adapter', () => {
 
   describe('#filesForVariant', () => {
     it('should return the variant file', () => {
-      const files = Adapter.filesForVariant('button', 'primary')
+      const files = Adapter.filesForVariant({}, 'button', 'primary')
 
       assert.strictEqual(files.length, 1)
       assert.strictEqual(files[0].basename, 'Primary.jsx')