+++ packages/cli/test/cli_test.js
@@ -37,7 +37,6 @@ describe('CLI', function () {
       assert.strictEqual(config.source.components, './src/components')
       assert.strictEqual(config.source.templates, './src/templates')
       assert.strictEqual(config.source.pages, './uiengine/pages')
-      assert.strictEqual(config.source.data, './uiengine/data')
       assert.strictEqual(config.source.entities, './uiengine/entities')
       assert.strictEqual(config.target, './dist')
       assert.strictEqual(config.adapters.html, '@uiengine/adapter-html')